### 2018-05-17 #740 @rowanc1
Thanks for the suggestion, cursory glance through things looks cool. :) Unfortunately, I won't be able to realistically look at this until late July. You might be able to find someone else, or I can aim for that time frame. Let me know!
