### 2018-05-18 #741 @whedon
[REVIEW]: PyUnfold: A Python Package for Iterative Unfolding
**Submitting author:** @zhampel (<a href="http://orcid.org/0000-0003-0253-9117">Zigfried Hampel-arias</a>)
**Repository:** <a href="https://github.com/jrbourbeau/pyunfold" target ="_blank">https://github.com/jrbourbeau/pyunfold</a>
**Version:** v0.4
**Archive:** <a href="https://doi.org/10.5281/zenodo.1258211">10.5281/zenodo.1258211</a>
