### 2018-05-18 #741 @lheagy
Many thanks @juliohm and @benjaminrose for being willing to review! In the main issue thread, there is a checklist for each of you as you go through the review. Please let me know if you have any questions or if there is anything I can assist with!
