### 2018-05-27 jrbourbeau/pyunfold#67 @juliohm
pyunfold not installed in binder tutorial
Hi @jrbourbeau,

I've started reviewing the package for JOSS. Could you please confirm that the binder tutorial should run just fine? I tried executing the first cell in the notebook, but it says that `pyunfold` is not installed.
