### 2018-05-27 #741 @juliohm
Very good documentation @zhampel, I was able to reproduce all the examples, and follow the mathematical derivations.

A few comments to improve the project further:

### Conceptual
- It would be nice to compare `pyunfold` with other related packages for inverse problems in Python, or clarify in the documentation the unique features of the project.
- Additionally, it would be interesting to see examples where the observed distribution is very different than the true distribution. Or alternatively, you could list the distributional assumptions under which the method is successful. For example, can the method be applied to update a bimodal observation into an unimodal truth?

### Implementation
- Priors (or distributions) in the project are represented with a vector of numbers (densities). In the section "custom priors" you could clarify that any vector of numbers (integrating to one) can be used as a prior and that priors `uniform` and `jeffreys` are not specialized classes.
- The binder option to run the examples online is missing the `pyunfold` module as reported in https://github.com/jrbourbeau/pyunfold/issues/67
- Since you are using Travis CI for the tests, you could also consider activating the build on Mac OS. You can do it by adding the section below:

```
os:
  - linux
  - osx
```
