### 2018-05-27 jrbourbeau/pyunfold#67 @jrbourbeau
Good catch @juliohm! I was able to reproduce the same error you got when running the binder tutorial

```python
ModuleNotFoundError: No module named 'pyunfold'
```

I'll update the environment that binder builds to make sure `pyunfold` is actually installed

Thanks for the JOSS review and isssue report!
