### 2018-05-27 c7fff5d
Fixes environment used in binder examples (#68)
* Adds binder/ directory

* Adds changelog entry
