### 2018-05-27 #741 @jrbourbeau
> The binder option to run the examples online is missing the pyunfold module

Thanks @juliohm for pointing this out and submitting a corresponding issue! I've updated the environment that is built to run the example notebooks on binder. `pyunfold` is now installed and the notebooks all work as expected.

ref: jrbourbeau/pyunfold#68
