### 2018-05-27 d2a2d3c
Adds osx tests to Travis CI
