+++ .travis.yml
@@ -8,6 +8,10 @@ python:
     - "3.5"
     - "3.6"
 
+os:
+    - linux
+    - osx
+
 matrix:
     fast_finish: true
 
