### 2018-05-27 427444f
Adds ci install.sh script
