+++ .travis.yml
@@ -1,41 +1,28 @@
-language: python
+language: generic
 
-cache: pip
+os: linux
 
-python:
-    - "2.7"
-    - "3.4"
-    - "3.5"
-    - "3.6"
+jobs:
+  fast_finish: true
+  include:
+    - env:
+      - PYTHON=2.7
 
-os:
-    - linux
-    - osx
+    - env:
+      - PYTHON=3.4
 
-matrix:
-    fast_finish: true
+    - env:
+      - PYTHON=3.5
+
+    - env:
+      - PYTHON=3.6
+
+    - env:
+      - PYTHON=3.6
+      os: osx
 
 install:
-    - sudo apt-get update
-    # We do this conditionally because it saves us some downloading if the
-    # version is the same.
-    - if [[ "$TRAVIS_PYTHON_VERSION" == "2.7" ]]; then
-        wget https://repo.continuum.io/miniconda/Miniconda2-latest-Linux-x86_64.sh -O miniconda.sh;
-      else
-        wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
-      fi
-    - bash miniconda.sh -b -p $HOME/miniconda
-    - export PATH="$HOME/miniconda/bin:$PATH"
-    - hash -r
-    - conda config --set always_yes yes --set changeps1 no
-    - conda update -q conda
-    # Useful for debugging any issues with conda
-    - conda info -a
-    - conda create -q --name test-environment python=$TRAVIS_PYTHON_VERSION
-    - conda env update --name=test-environment --file=environment.yml
-    - source activate test-environment
-    # Install pyunfold + dev dependencies
-    - pip install .[dev]
+    - source ci/install.sh;
 
 script:
     - pytest --cov=pyunfold pyunfold
@@ -58,4 +45,4 @@ deploy:
     email: "github-pages-deploy@travis-ci.org"
     on:
         branch: master
-        python: "3.6"
+        condition: $PYTHON = 3.6 && $TRAVIS_OS_NAME = linux
