+++ ci/after_success.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+if [[ $COVERAGE == 'true' ]]; then
+    pip install codecov
+    codecov
+fi
