### 2018-05-27 ae9afbf
Adds DOCS variable
