### 2018-05-27 aaebf2f
Updates deploy condition
