### 2018-05-27 71b1e2c
Adds osx tests to Travis CI (#69)
* Adds osx tests to Travis CI

* Adds ci install.sh script

* Adds after_success.sh

* Adds DOCS variable

* Updates deploy condition
