### 2018-05-27 #741 @jrbourbeau
> Since you are using Travis CI for the tests, you could also consider activating the build on Mac OS.

That's a good point. I added Mac OS tests for Python 3.6 in jrbourbeau/pyunfold#69.

I didn't want to double the number of jobs that are run on Travis by running both `linux` and `osx` tests for several different versions of Python. So instead I'm keeping the original jobs that run Python 2.7, 3.4, 3.5, and 3.6 on `linux` and have added a job for Python 3.6 on `osx`. This should catch if there is a Python-specific problem or a Mac OS-specific problem.
