### 2018-05-27 #741 @juliohm
Thank you @jrbourbeau for incorporating the changes. Is there a good reason to test on all 3.x versions of Python? Are they very different?

I will convert my previous comment into a set of tasks, and will mark the tasks you've already implemented as done. :heavy_check_mark:
