### 2018-05-27 #741 @jrbourbeau
> Is there a good reason to test on all 3.x versions of Python? Are they very different?

All of the existing code in PyUnfold should work just fine on all Python 3.x versions. Running the automated tests on different 3.x versions is more a preventative measure to make sure new contributions don't unexpectedly add something that isn't fully backwards compatible (e.g. Python 3.6 added a new [f-string formatting syntax](https://docs.python.org/3/reference/lexical_analysis.html#f-strings) which are really nice, but won't work on Python 3.4 or 3.5).
