### 2018-05-28 14a4bc1
Adds priors API documentation
