### 2018-05-28 2100fc2
JOSS review-Clarify custom prior documentation (#70)
* Adds priors API documentation

* Updates custom priors documentation
