### 2018-05-28 #741 @jrbourbeau
> you could clarify that any vector of numbers (integrating to one) can be used as a prior and that priors uniform and jeffreys are not specialized classes

@juliohm I've updated the [Custom Priors section](https://jrbourbeau.github.io/pyunfold/notebooks/user_prior.html#Custom-Priors) in the documentation in jrbourbeau/pyunfold#70 to address the points you brought up. Hopefully it is more clear now that any given `prior` array can be used, not just the built-in `uniform_prior` or `jeffreys_prior`. Please let me know if you think it needs any further clarification.

Thank you for the thorough read through and feedback!
