### 2018-05-28 e2d747c
Correcting bullet list
