### 2018-05-29 f4f8bd3
Adding non-negative value check for input data objects
