### 2018-05-29 cf5ff87
any comparison to float
