### 2018-05-29 638c431
Using array to ensure any comparison succeeds
