### 2018-05-29 1176341
Adding unit test for non-negativity
