### 2018-05-29 a49a18b
Adds NSF to JOSS paper acknowledgements
