### 2018-05-29 6c2e8b2
Adds NSF to JOSS paper acknowledgements (#75)
* Adds NSF to JOSS paper acknowledgements

* Fixes typo
