### 2018-05-29 edad533
Removes whitespace [skip ci]
