### 2018-05-29 9f8921d
JOSS review-Add why pyunfold page (#71)
* Adding why pyunfold tab

* Correcting bullet list

* Edits to why pyunfold page

* Removes whitespace [skip ci]
