+++ docs/source/index.rst
@@ -61,6 +61,7 @@ PyUnfold provides an unfolding toolkit for members of all scientific disciplines
     :caption: Getting Started
 
     overview
+    why
     installation
     Tutorial <notebooks/tutorial.ipynb>
 
