### 2018-05-29 #741 @jrbourbeau
> clarify in the documentation the unique features of the project

@juliohm we've added a ["Why PyUnfold?" ](https://jrbourbeau.github.io/pyunfold/why) page to the documentation that outlines the unique features of the project.

ref: jrbourbeau/pyunfold#71
