### 2018-05-30 1e0d31f
Updates negative value check
