### 2018-05-30 ab08e2d
Adds changelog entry
