### 2018-05-30 5f33444
Adding non-negative value check for input data objects (#73)
* Adding non-negative value check for input data objects

* any comparison to float

* Using array to ensure any comparison succeeds

* Adding unit test for non-negativity

* Updates negative value check

* Adds changelog entry
