+++ pyunfold/unfold.py
@@ -120,7 +120,9 @@ def iterative_unfold(data=None, data_err=None, response=None,
               }
     for name in inputs:
         if inputs[name] is None:
-            raise ValueError('The input for "{}" must not be None.'.format(name))
+            raise ValueError('The input for {} must not be None.'.format(name))
+        elif np.amin(inputs[name]) < 0:
+            raise ValueError('The items in {} must be non-negative.'.format(name))
 
     data, data_err = cast_to_array(data, data_err)
     response, response_err = cast_to_array(response, response_err)
