### 2018-05-30 #741 @jrbourbeau
> Additionally, it would be interesting to see examples where the observed distribution is very different than the true distribution...For example, can the method be applied to update a bimodal observation into an unimodal truth?

@juliohm Indeed a bimodal observed distribution can be unfolded into a unimodal truth distribution. For example, in a similar fashion to the "Tutorial" notebook, one can have true, observed, and unfolded distributions as shown below.

![image](https://user-images.githubusercontent.com/11656932/40750330-8af0e36c-642c-11e8-81e3-43fa1326b81b.png)

I can add this as an additional example notebook to the documentation. My only concern is that it wouldn't really walk through any new functionality in PyUnfold, just simply using a different example dataset.
