### 2018-05-30 #741 @juliohm
That is a very nice example, I think it tells a lot more about the generality of the method than the current example in the documentation, but it is up to you to decide what example is more telling for the audience you are targeting. =)

From a user's perspective, or maybe putting myself in the mind of someone that is probing different packages for inverse problems, I'd say that just knowing that the method doesn't assume gaussianity or bell shapes is a big plus. :+1:

I am marking all the items as solved. Congratulations again on the package, good stuff!

@benjaminrose do you have further corrections or improvements that you would like to share?
