### 2018-05-30 #741 @jrbourbeau
> just knowing that the method doesn't assume gaussianity or bell shapes is a big plus

That's a really good point. I'll discuss with @zhampel as to which example dataset would be more illuminating for the unfolding community.
