### 2018-05-31 #741 @zhampel
@juliohm so we've used the method on power law spectra in publications, so @jrbourbeau and I can cook up an example from simulation for example that demonstrates a non-gaussian data set. We just have to decide how to make our fake data without violating our collaborations's respective public release agreements ;)
