### 2018-05-31 #741 @zhampel
@juliohm we've used the method successfully on power law spectra unfoldings for publications, so @jrbourbeau and I can cook up such an example in addition to our canonical examples. We just have to make sure we don't violate any public release agreements from our respective collaborations ;)
