### 2018-05-31 #741 @benjaminrose
This is a great project. @juliohm's thank you for your great comments. A have a few of my own to add.

# Docs - Tutorial

- [ ] Binder is a great tool for this! Thanks for using it. Also, maybe add a line about needing to install `matplotlib`  if you want to follow this tutorial on your own.
- [ ] Can you point out the cause for "We can see how the structure of the 2-d response histogram differs from the normalized response matrix."
- [ ] In the tutorial, can you either add a short explanation of the `ts` parameter is or just use the default values for input 20.

# Docs - API
- [ ] Can you define the `ts` stats in the API docs, only their acronyms are used.
    - The paper does define them, but not the API docs
- [ ] Can you explain/link to a definition of Jeffrey's priors in the API docs? Maybe as a References header?

# Docs - Contributing page
- [ ] The how to contribute guide is great. Like you liked to Github's article on forks, maybe in step 11 add a link to Github's recommendation to writing a pull request, https://blog.github.com/2015-01-21-how-to-write-the-perfect-pull-request/.
* Thanks for explaining `git fetch upstream`. This is very useful for first-time contributors.

# Paper

- [ ] The statement of need is better in the docs than in the paper. I really like [this section](https://jrbourbeau.github.io/pyunfold/overview.html#what-is-unfolding).

# General

* What is your plan to get to v1.0?
* Walking through one simple but complete example is enough. I support the idea of at least pointing out more complicated solutions are solve-able, like the bimodal observed distribution coming from a unimodal truth. Putting it with the other advanced techniques makes sense to me.
