### 2018-05-31 #741 @jrbourbeau
Thanks for your review @benjaminrose! I'll get started on addressing your comments and link back to this review in the relevant PRs.
