### 2018-05-31 85cb699
Updates teststat module
