+++ pyunfold/teststat.py
@@ -1,9 +1,7 @@
 
 from __future__ import division, print_function
 import numpy as np
-from scipy.special import gammainc as gammaq
 from scipy.special import gammaln as lgamma
-from scipy.stats import kstwobign
 
 from .utils import none_to_empty_list
 
@@ -11,61 +9,49 @@ from .utils import none_to_empty_list
 class TestStat(object):
     """Common base class for test statistic methods
     """
-    def __init__(self, name="TestStat", tol=None, num_causes=None,
-                 TestRange=None, verbose=False, **kwargs):
-        TestRange = none_to_empty_list(TestRange)
-        # TS Name
-        self.name = name
-        # User Defined TS Tolerance
+    def __init__(self, tol=None, num_causes=None, test_range=None, **kwargs):
+        test_range = none_to_empty_list(test_range)
         self.tol = tol
         if num_causes is None:
             raise ValueError('Number of causes (num_causes) must be provided.')
         cause_bin_edges = np.arange(num_causes + 1, dtype=float)
         # Get bin midpoints
         cause_axis = (cause_bin_edges[1:] + cause_bin_edges[:-1]) / 2
-        self.Xaxis = cause_axis
-        self.TSRange = TestRange
-        self.IsRangeTS = False
-        self.TSbins = self.SetTestRangeBins()
-        # Verbose Functionality
-        self.verbose = verbose
-        self.printProbMessage = True
-        self.printStatsHeader = True
-        if verbose:
-            self.PrintName()
+        self.cause_axis = cause_axis
+        self.ts_range = test_range
+        self.has_ts_range = False
+        self.ts_bins = self.set_test_range_bins()
         # Initialize Unnatural TS data
         self.stat = np.inf
-        self.delstat = 0
-        self.prob = -1
         self.dof = -1
         self.dofSet = False
 
-    def SetTestRangeBins(self):
+    def set_test_range_bins(self):
         bins = [0, -1]
-        if (not self.TSRange == []):
-            lTSR = len(self.TSRange)
+        if self.ts_range != []:
+            lTSR = len(self.ts_range)
             err_mess = ("***\n Test stat range can only have two elements. "
                         "This has {}. Exiting...***\n".format(lTSR))
             assert lTSR == 2, err_mess
-            xlo = self.TSRange[0]
-            xhi = self.TSRange[1]
+            xlo = self.ts_range[0]
+            xhi = self.ts_range[1]
             err_mess = "***\n Test stat limits reversed. xlo must be < xhi. Exiting...***\n"
             assert xlo < xhi, err_mess
 
             # Find the bins corresponding to the test range requested
-            lobin = np.searchsorted(self.Xaxis, xlo)
-            hibin = np.searchsorted(self.Xaxis, xhi)
+            lobin = np.searchsorted(self.cause_axis, xlo)
+            hibin = np.searchsorted(self.cause_axis, xhi)
             bins = [lobin, hibin]
-            self.IsRangeTS = True
+            self.has_ts_range = True
         return bins
 
-    def GetArrayRange(self, N1, N2):
-        if self.IsRangeTS:
-            NR1 = N1[self.TSbins[0]:self.TSbins[1]]
-            NR2 = N2[self.TSbins[0]:self.TSbins[1]]
+    def get_array_range(self, dist1, dist2):
+        if self.has_ts_range:
+            NR1 = dist1[self.ts_bins[0]:self.ts_bins[1]]
+            NR2 = dist2[self.ts_bins[0]:self.ts_bins[1]]
             return NR1, NR2
         else:
[…]
