### 2018-05-31 95d013f
Adds Jeffreys link to API docs
