### 2018-05-31 5b71f0b
JOSS review-Update API documentation (#77)
* Updates teststat module

* Adds test statistics to API docs

* Adds Jeffreys link to API docs
