### 2018-05-31 adfa292
Adds link to PR best practices articlfe
