### 2018-05-31 #741 @jrbourbeau
@benjaminrose At this point I think I've addressed your comments related to the tutorial notebook, API documentation, and contributing guide. Please let me know if further clarification / changes are needed to these sections of the documentation.

Also, thanks for the link to the "How to write the perfect pull request" article! There's a bunch of really good stuff in there.
