### 2018-05-31 #741 @jrbourbeau
> The statement of need is better in the docs than in the paper

@benjaminrose Just to clarify, are you recommending we include the content from the "What is unfolding?" documentation in the paper?
