### 2018-05-31 d6cf5c1
Adds github issue/PR templates and CONTRIBUTING.md (#79)
