+++ .github/CONTRIBUTING.md
@@ -1,0 +1,1 @@
+If you would like to contribute to PyUnfold, see our [contributing guide](https://jrbourbeau.github.io/pyunfold/contributing.html)!
