### 2018-06-01 4a9ea94
Adding et al. for cosmic ray paper ref (#80)
