+++ docs/joss_paper/paper.bib
@@ -86,7 +86,7 @@
 
 @article{hawc-crspectrum,
   title = {All-particle cosmic ray energy spectrum measured by the {HAWC} experiment from 10 to 500 TeV},
-  author = {{Alfaro, R.} and others},
+  author = {{Alfaro, R.} et al.},
   collaboration = {HAWC Collaboration},
   journal = {Phys. Rev. D},
   volume = {96},
