+++ docs/joss_paper/paper.bib
@@ -86,7 +86,7 @@
 
 @article{hawc-crspectrum,
   title = {All-particle cosmic ray energy spectrum measured by the {HAWC} experiment from 10 to 500 TeV},
-  author = {{Alfaro, R.} et al.},
+  author = {Alfaro, R. and Alvarez, C. and \'Alvarez, J. D. and Arceo, R. and Arteaga-Vel\'azquez, J. C. and Avila Rojas, D. and Ayala Solares, H. A. and Barber, A. S. and Becerril, A. and Belmont-Moreno, E. and BenZvi, S. Y. and Brisbois, C. and Caballero-Mora, K. S. and Capistr\'an, T. and Carrami\~nana, A. and Casanova, S. and Castillo, M. and Cotti, U. and Cotzomi, J. and Couti\~no de Le\'on, S. and De Le\'on, C. and De la Fuente, E. and Diaz Hernandez, R. and Dichiara, S. and Dingus, B. L. and DuVernois, M. A. and D\'{\i}az-V\'elez, J. C. and Ellsworth, R. W. and Enriquez-Rivera, O. and Fiorino, D. W. and Fleischhack, H. and Fraija, N. and Garc\'{\i}a-Gonz\'alez, J. A. and Gonz\'alez Mu\~noz, A. and Gonz\'alez, M. M. and Goodman, J. A. and Hampel-Arias, Z. and Harding, J. P. and Hernandez-Almada, A. and Hinton, J. and Hueyotl-Zahuantitla, F. and Hui, C. M. and H\"untemeyer, P. and Iriarte, A. and Jardin-Blicq, A. and Joshi, V. and Kaufmann, S. and Lara, A. and Lauer, R. J. and Lennarz, D. and Le\'on Vargas, H. and Linnemann, J. T. and Longinotti, A. L. and Luis Raya, G. and Luna-Garc\'{\i}a, R. and L\'opez-C\'amara, D. and L\'opez-Coto, R. and Malone, K. and Marinelli, S. S. and Martinez, O. and Martinez-Castellanos, I. and Mart\'{\i}nez-Castro, J. and Mart\'{\i}nez-Huerta, H. and Matthews, J. A. and Miranda-Romagnoli, P. and Moreno, E. and Mostaf\'a, M. and Nellen, L. and Newbold, M. and Nisa, M. U. and Noriega-Papaqui, R. and Pelayo, R. and Pretz, J. and P\'erez-P\'erez, E. G. and Ren, Z. and Rho, C. D. and Rivi\`ere, C. and Rosa-Gonz\'alez, D. and Rosenberg, M. and Ruiz-Velasco, E. and Salesa Greus, F. and Sandoval, A. and Schneider, M. and Schoorlemmer, H. and Sinnis, G. and Smith, A. J. and Springer, R. W. and Surajbali, P. and Taboada, I. and Tibolla, O. and Tollefson, K. and Torres, I. and Ukwatta, T. N. and Villase\~nor, L. and Weisgarber, T. and Westerhoff, S. and Wood, J. and Yapici, T. and Zepeda, A. and Zhou, H.},
   collaboration = {HAWC Collaboration},
   journal = {Phys. Rev. D},
   volume = {96},
