### 2018-06-01 #741 @benjaminrose
@jrbourbeau, I think the "What is unfolding?" is your strongest statement of need. The paper is fine, but I feel like you have a better argument somewhere else. Having the best argument in the docs is helpful since most people will read that and not the paper. If you see a simple way to add it, and you want to, I think it would make the paper better.
