### 2018-06-01 #741 @benjaminrose
The other changes look great @jrbourbeau. But I missed the API docs for the `teststat` class.   The abbreviations that are the not fully that I was talking about were in describing the `ts` parameter for `iterative_unfold()`. I would just add a "more info" link to https://jrbourbeau.github.io/pyunfold/api.html#test-statistics at the end of that parameter's description.
