### 2018-06-01 #741 @jrbourbeau
> I was talking about were in describing the ts parameter for iterative_unfold()

Gotcha, I misunderstood your comment. Thanks for noticing that! I'll update the `iterative_unfold` docstring accordingly
