### 2018-06-01 bb6f015
Updates ts parameter documentation
