### 2018-06-01 #741 @jrbourbeau
@benjaminrose I updated the `ts` parameter description with a link to the test statistics section in the API docs
