### 2018-06-01 #741 @jrbourbeau
@juliohm just as an FYI the "Performance" checkbox under the "Functionality" section for your review checklist is unchecked. If there's anything else that needs to be done on my end, just let me know. Thanks!
