### 2018-06-01 #741 @juliohm
@jrbourbeau I don't recall reading in the docs statements about performance? Please let me know if they exist somewhere. The checkbox should only be marked in case the software claims states explicitly that it is "high-performance" or faster than alternatives.
