### 2018-06-01 #741 @benjaminrose
@lheagy, I think  @zhampel  and @jrbourbeau have a great project here and I fully endorse their paper for publication in JOSS.
