### 2018-06-01 ef7a8fe
Adds section from docs to paper
