### 2018-06-01 01dbeee
JOSS review-Update statement of need in paper  (#83)
* Adds section from docs to paper

* Updates HAWC citation formatting
