+++ docs/joss_paper/paper.bib
@@ -81,12 +81,12 @@
   doi             = {10.1088/0004-637X/738/1/82},
   url             = {http://stacks.iop.org/0004-637X/738/i=1/a=82},
   year            = {2011},
-}	
+}
 
 
 @article{hawc-crspectrum,
   title = {All-particle cosmic ray energy spectrum measured by the {HAWC} experiment from 10 to 500 TeV},
-  author = {{Alfaro, R.} et al.},
+  author = {Alfaro, R. et al.},
   collaboration = {HAWC Collaboration},
   journal = {Phys. Rev. D},
   volume = {96},
