### 2018-06-01 #741 @jrbourbeau
>  I think the "What is unfolding?" is your strongest statement of need...If you see a simple way to add it, and you want to, I think it would make the paper better.

@benjaminrose I agree, that's a really good point. I've added the content from the "What is unfolding?" docs to the beginning of the paper. I think it helps clarify the problem we're addressing with PyUnfold without messes with the overall flow of the paper. Thanks for the feedback!

ref: jrbourbeau/pyunfold#83
