### 2018-06-01 #741 @jrbourbeau
> The checkbox should only be marked in case the software claims states explicitly that it is "high-performance" or faster than alternatives.

@juliohm You're correct, we don't make any statements about the performance in the docs. However, I interpreted the "If there are no claims, please check off this item" sentence to mean that since we make no performance claims, then the performance checkbox should be marked as done. I could definitely be misinterpreting that though.
