### 2018-06-01 #741 @juliohm
Oh maybe it is a language issue, I am not a native speaker. When I read "Check off" I interpret it as "Do not mark the box", it is probably the opposite?
