### 2018-06-01 #741 @jrbourbeau
I took "check off" to mean "mark the box". Perhaps @lheagy can clarify this point for us.
