### 2018-06-03 #741 @lheagy
Hi @juliohm and @jrbourbeau: sorry for the confusion (and I appreciate knowing that this is a potential point of confusion for non-native speakers). By "check off", we do mean that you can "mark the box" if there are no claims made.
