### 2018-06-03 #741 @jrbourbeau
@lheagy at this point both the reviewers have approved the paper / project for publication in JOSS. Are there any additional steps I need to take in order to move forward with publication? Thanks!
