### 2018-06-03 #741 @lheagy
Excellent! Thank you @juliohm and @benjaminrose for your review!!

@zhampel and @jrbourbeau at this point, could you please archive your software on [zenodo](https://zenodo.org/) or similar and post the doi here?
