### 2018-06-03 fb660fc
Adds changelog entry for JOSS review changes
