### 2018-06-03 8c1f73e
Bumps version number to 0.4
