+++ docs/source/changelog.rst
@@ -6,8 +6,8 @@
 Release Notes
 *************
 
-Version 0.4.dev0 (TBD)
-----------------------
+Version 0.4 (2018-06-03)
+------------------------
 
 **Changes**:
 
@@ -15,7 +15,7 @@ Version 0.4.dev0 (TBD)
   distributions. An error is raised if negative efficiencies, responses, or
   observed data distributions are input.
   (See `PR #73 <https://github.com/jrbourbeau/pyunfold/pull/73>`_)
-- Adds several documentation updates in response to JOSS reviewer comments 
+- Adds several documentation updates in response to JOSS reviewer comments
   (See
   `PR #83 <https://github.com/jrbourbeau/pyunfold/pull/83>`_,
   `PR #82 <https://github.com/jrbourbeau/pyunfold/pull/82>`_,
