### 2018-06-03 d3cc41a
V0.4 release (#84)
* Adds changelog entry for JOSS review changes

* Bumps version number to 0.4
