+++ pyunfold/__version__.py
@@ -1,2 +1,2 @@
 
-__version__ = '0.4.dev0'
+__version__ = '0.4'
