### 2018-06-03 #741 @lheagy
Thanks @zhampel: quick question for you - it looks like the version you archived is v0.4. This is the version we should list with the publication?
