### 2018-06-04 #741 @zhampel
Thank you so much @lheagy @juliohm @benjaminrose for your thorough and speedy review! 🎉 We look forward to further contributions to JOSS.
